+++ src/main/java/com/microservicesteam/nutaxi/infrastructure/CacheConfiguration.java
@@ -1,0 +1,17 @@
+package com.microservicesteam.nutaxi.infrastructure;
+
+import org.springframework.cache.annotation.CachingConfigurerSupport;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.cache.interceptor.CacheErrorHandler;
+import org.springframework.context.annotation.Configuration;
+
+@EnableCaching
+@Configuration
+public class CacheConfiguration extends CachingConfigurerSupport {
+
+    @Override
+    public CacheErrorHandler errorHandler() {
+        return new DefaultCacheErrorHandler();
+    }
+
+}
