+++ src/main/java/com/microservicesteam/nutaxi/infrastructure/CloudConfiguration.java
@@ -1,0 +1,30 @@
+package com.microservicesteam.nutaxi.infrastructure;
+
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.commons.util.InetUtils;
+import org.springframework.cloud.netflix.eureka.EurekaInstanceConfigBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+
+import com.netflix.appinfo.AmazonInfo;
+
+@EnableDiscoveryClient
+@Configuration
+public class CloudConfiguration {
+
+    @Bean
+    @Profile("docker-aws")
+    public EurekaInstanceConfigBean eurekaInstanceConfig(InetUtils inetUtils, EmbeddedServletContainerRandomPortListener portListener) {
+        EurekaInstanceConfigBean config = new EurekaInstanceConfigBean(inetUtils);
+        AmazonInfo info = AmazonInfo.Builder.newBuilder().autoBuild("eureka");
+        config.setDataCenterInfo(info);
+        info.getMetadata().put(AmazonInfo.MetaDataKey.publicHostname.getName(), info.get(AmazonInfo.MetaDataKey.publicIpv4));
+        config.setInstanceId(info.get(AmazonInfo.MetaDataKey.instanceId));
+        config.setHostname(info.get(AmazonInfo.MetaDataKey.publicHostname));
+        config.setIpAddress(info.get(AmazonInfo.MetaDataKey.publicIpv4));
+        config.setNonSecurePort(portListener.getPort());
+        return config;
+    }
+
+}
