+++ src/test/java/com/microservicesteam/nutaxi/infrastructure/CloudConfigurationTest.java
@@ -1,4 +1,4 @@
-package com.microservicesteam.nutaxi;
+package com.microservicesteam.nutaxi.infrastructure;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
@@ -11,12 +11,9 @@ import org.springframework.cloud.commons.util.InetUtils;
 import org.springframework.cloud.commons.util.InetUtilsProperties;
 import org.springframework.cloud.netflix.eureka.EurekaInstanceConfigBean;
 
-import com.microservicesteam.nutaxi.infrastructure.EmbeddedServletContainerRandomPortListener;
-
 @RunWith(MockitoJUnitRunner.class)
-public class NutaxiRouteServiceApplicationTest {
-
-    private NutaxiRouteServiceApplication underTest;
+public class CloudConfigurationTest {
+    private CloudConfiguration underTest;
 
     @Mock
     private EmbeddedServletContainerRandomPortListener portListener;
@@ -25,7 +22,7 @@ public class NutaxiRouteServiceApplicationTest {
     public void shouldConfigureEurekaInstanceConfigBean() {
         when(portListener.getPort()).thenReturn(12345);
 
-        underTest = new NutaxiRouteServiceApplication();
+        underTest = new CloudConfiguration();
 
         EurekaInstanceConfigBean eurekaInstanceConfig = underTest.eurekaInstanceConfig(getINetUtils(), portListener);
